feat(user): resolve register question
For users without any register question and without any register event,
find the very first session, and try to find a question
from the events of this very first session.

If one is found, use it as a register question for the user.

These users are  mostly "old" users at the time when the register event
was not emitted.

Signed-off-by: Francois LAROCHE <fl@make.org>