chore(transverse): move to grizzled-slf4j
This migration keeps the `logger.` before the logging function and
doesn't use the info, error, warn, etc.. functions to avoid name
collisions when there are variables called, for instance, error.

When it's the case, it seems like the cmpiler has troubles understanding
whether it should consider this variable as an error or use the function
defined in the trait.

Signed-off-by: Francois LAROCHE <fl@make.org>